+++ NEWS.md
@@ -1,6 +1,9 @@
 # dplyr (development version)
 
+* `select()` not creating duplicate variable with a grouped data frame (#5841).
+
 * Fixed behaviour of `...` in top-level `across()` calls (#5813, #5832).
+
 * `dplyr_col_select()` keeps attributes for bare data frames (#5294, #5831).
 
 * Fixed quosure handling in `dplyr::group_by()` that caused issues with extra